refactor(types): move non-core, non-public types out of the types package
