+++ packages/docusaurus-mdx-loader/src/deps.d.ts
@@ -7,10 +7,13 @@
 
 // TODO Types provided by MDX 2.0 https://github.com/mdx-js/mdx/blob/main/packages/mdx/types/index.d.ts
 declare module '@mdx-js/mdx' {
-  import type {Processor} from 'unified';
-  import type {MDXPlugin} from '@docusaurus/mdx-loader';
+  import type {Processor, Plugin} from 'unified';
 
-  export type Options = {
+  type MDXPlugin =
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    [Plugin<any[]>, any] | Plugin<any[]>;
+
+  type Options = {
     filepath?: string;
     skipExport?: boolean;
     wrapExport?: string;
