+++ packages/docusaurus-mdx-loader/tsconfig.json
@@ -7,5 +7,6 @@
     "declarationMap": true,
     "rootDir": "src",
     "outDir": "lib"
-  }
+  },
+  "include": ["src"]
 }
