+++ packages/docusaurus-plugin-content-docs/src/plugin-content-docs.d.ts
@@ -7,8 +7,13 @@
 
 declare module '@docusaurus/plugin-content-docs' {
   import type {MDXOptions} from '@docusaurus/mdx-loader';
-  import type {ContentPaths, FrontMatterTag} from '@docusaurus/utils';
-  import type {TagsListItem, TagModule, Tag} from '@docusaurus/types';
+  import type {
+    ContentPaths,
+    FrontMatterTag,
+    TagsListItem,
+    TagModule,
+    Tag,
+  } from '@docusaurus/utils';
   import type {Required} from 'utility-types';
 
   export type Assets = {
