+++ packages/docusaurus-utils-validation/src/validationSchemas.ts
@@ -6,8 +6,7 @@
  */
 
 import Joi from './Joi';
-import {isValidPathname, DEFAULT_PLUGIN_ID} from '@docusaurus/utils';
-import type {Tag} from '@docusaurus/types';
+import {isValidPathname, DEFAULT_PLUGIN_ID, type Tag} from '@docusaurus/utils';
 import {JoiFrontMatter} from './JoiFrontMatter';
 
 export const PluginIdSchema = Joi.string()
